tools/giftomp4.py
#!/usr/bin/env python3
"""
GIF to MP4 converter tool for Jackknife.

This tool converts GIF animations to MP4 video files using Pillow and imageio.
"""

import argparse
import sys
import time
from pathlib import Path

# Try importing dependencies
try:
    from PIL import Image

    PILLOW_VERSION = Image.__version__
except ImportError:
    PILLOW_VERSION = "NOT FOUND - Environment may not be set up correctly"

# Try importing Rich for better output (if available)
try:
    from rich.console import Console
    from rich.progress import Progress, BarColumn, TextColumn
    from rich.panel import Panel

    console = Console()
    RICH_AVAILABLE = True
except ImportError:
    RICH_AVAILABLE = False


def convert_gif_to_mp4(input_path: str, output_path: str, fps: int) -> None:
    """
    Convert a GIF animation to MP4 video format.

    This is currently a placeholder simulation.
    """
    if RICH_AVAILABLE:
        console.print(
            f"Converting '[bold cyan]{input_path}[/]' to '[bold green]{output_path}[/]' at [bold]{fps}[/] FPS"
        )
        console.print(f"Using Pillow version: [bold yellow]{PILLOW_VERSION}[/]")
    else:
        print(f"Converting '{input_path}' to '{output_path}' at {fps} FPS")
        print(f"Using Pillow version: {PILLOW_VERSION}")

    # Example using Pillow to open and analyze the GIF
    try:
        with Image.open(input_path) as im:
            if RICH_AVAILABLE:
                console.print(
                    f"[green]Successfully opened[/] '{input_path}' with Pillow"
                )
                console.print(
                    f"Image format: [cyan]{im.format}[/], Size: [cyan]{im.size}[/], Mode: [cyan]{im.mode}[/]"
                )

                if getattr(im, "is_animated", False):
                    console.print(
                        f"Detected [bold cyan]{getattr(im, 'n_frames', 1)}[/] frames"
                    )
                else:
                    console.print(
                        "[yellow]Warning:[/] Input might not be an animated GIF"
                    )
            else:
                print(f"Successfully opened '{input_path}' with Pillow")
                print(f"Image format: {im.format}, Size: {im.size}, Mode: {im.mode}")

                if getattr(im, "is_animated", False):
                    print(f"Detected {getattr(im, 'n_frames', 1)} frames")
                else:
                    print("Warning: Input might not be an animated GIF")
    except Exception as e:
        if RICH_AVAILABLE:
            console.print(f"[bold red]Error processing image:[/] {e}")
        else:
            print(f"Error processing image: {e}")

    # Simulate conversion process
    if RICH_AVAILABLE:
        console.print("Processing frames...")
        with Progress(
            TextColumn("[progress.description]{task.description}"),
            BarColumn(),
            TextColumn("[progress.percentage]{task.percentage:>3.0f}%"),
        ) as progress:
            task = progress.add_task("Converting", total=100)
            for i in range(10):
                time.sleep(0.3)  # Simulate work
                progress.update(task, advance=10)
    else:
        print("Processing frames...")
        for i in range(5):
            time.sleep(0.3)  # Simulate work
            print(f"Progress: {(i + 1) * 20}%")

    # Create a dummy output file for demo purposes
    if output_path and not Path(output_path).exists():
        try:
            Path(output_path).touch()
            if RICH_AVAILABLE:
                console.print(f"[green]Created output file:[/] {output_path}")
            else:
                print(f"Created output file: {output_path}")
        except Exception as e:
            if RICH_AVAILABLE:
                console.print(f"[bold red]Error creating output file:[/] {e}")
            else:
                print(f"Error creating output file: {e}")

    if RICH_AVAILABLE:
        console.print(Panel("Conversion complete!", style="green"))
    else:
        print("Conversion complete!")


def main() -> None:
    """Main entry point for the tool."""
    parser = argparse.ArgumentParser(
        description="giftomp4: Convert GIF animations to MP4 videos"
    )
    parser.add_argument("input_gif", help="Path to the input GIF file")
    parser.add_argument(
        "-o",
        "--output",
        help="Path for the output MP4 file (defaults to input filename with .mp4 extension)",
    )
    parser.add_argument(
        "--fps",
        type=int,
        default=15,
        help="Frames per second for the output video (default: 15)",
    )
    parser.add_argument(
        "-v", "--verbose", action="store_true", help="Enable verbose output"
    )

    args = parser.parse_args()

    if args.verbose:
        if RICH_AVAILABLE:
            console.print(Panel("giftomp4 Tool - Verbose Mode", style="blue"))
            console.print(f"Arguments: {args}")
            console.print("-----------------------------------")
        else:
            print("giftomp4 Tool - Verbose Mode")
            print(f"Arguments: {args}")
            print("-----------------------------------")

    # Determine output filename if not specified
    output_file = args.output or (Path(args.input_gif).stem + ".mp4")

    # Run the conversion
    convert_gif_to_mp4(args.input_gif, output_file, args.fps)


if __name__ == "__main__":
    main()
